Move alias to models/record.py
