+++ src/backend/expungeservice/endpoints/search.py
@@ -8,8 +8,8 @@ import logging
 
 from expungeservice.models.ambiguous import AmbiguousRecord
 from expungeservice.record_merger import RecordMerger
-from expungeservice.models.record import Question
-from expungeservice.record_creator import RecordCreator, Alias
+from expungeservice.models.record import Question, Alias
+from expungeservice.record_creator import RecordCreator
 from expungeservice.request import check_data_fields
 from expungeservice.request import error
 from expungeservice.serializer import ExpungeModelEncoder
